fix(worker-sessions): clear CI contract and lint gates
