+++ pkg/services/worker_sessions/transports/http/handler_test.go
@@ -180,14 +180,6 @@ func assertListObservationTurn(t *testing.T, observation factoryapi.WorkerSessio
 	}
 }
 
-func assertListObservationExecutionFacts(t *testing.T, observation factoryapi.WorkerSessionObservation) {
-	t.Helper()
-	if observation.Model == nil || *observation.Model != "gpt-5.6-luna" ||
-		observation.ReasoningEffort == nil || *observation.ReasoningEffort != "high" {
-		t.Fatalf("execution facts = model:%#v reasoningEffort:%#v, want gpt-5.6-luna/high", observation.Model, observation.ReasoningEffort)
-	}
-}
-
 func TestListWorkerSessionsBySessionIDReturnsEmptyForKnownWorkWithoutObservations(t *testing.T) {
 	service := &fakeObservationService{listErr: workersessions.ErrObservationWorkNotFound}
 	handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
@@ -404,14 +396,6 @@ func assertFailureObservationParse(t *testing.T, response factoryapi.WorkerSessi
 	t.Helper()
 	if response.Parse.EventCount != 4 || len(response.Parse.Errors) != 1 {
 		t.Fatalf("parse = %#v, want event and parse diagnostics", response.Parse)
-	}
-}
-
-func assertFailureObservationExecutionFacts(t *testing.T, response factoryapi.WorkerSessionObservation) {
-	t.Helper()
-	if response.Model == nil || *response.Model != "gpt-5.6-luna" ||
-		response.ReasoningEffort == nil || *response.ReasoningEffort != "medium" {
-		t.Fatalf("execution facts = model:%#v reasoningEffort:%#v, want gpt-5.6-luna/medium", response.Model, response.ReasoningEffort)
 	}
 }
